Moved SectionEntry to be a module
