+++ plugin-hrm-form/src/components/SectionEntry/index.ts
@@ -1,0 +1,1 @@
+export { default as SectionEntry } from './SectionEntry';
