+++ plugin-hrm-form/src/components/common/forms/Section.tsx
@@ -8,7 +8,7 @@ import {
   SectionTitleText,
   SectionCollapse,
   ContactDetailsIcon,
-} from '../styles/search';
+} from '../../../styles/search';
 
 const ArrowDownIcon = ContactDetailsIcon(ArrowDropDownTwoTone);
 const ArrowUpIcon = ContactDetailsIcon(ArrowDropUpTwoTone);
